[feature #146291077] refactor and group similar components into folders
