+++ client/components/routes.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Route, IndexRoute } from 'react-router';
-import Main from './main';
-import Login from './login';
-import Dashboard from './dashboard';
+import Dashboard from './common/dashboard';
+import Main from './common/main';
+import Login from './users/login';
 
 /**
  * Routes
